+++ lib/constants/ble_constants.dart
@@ -1,0 +1,10 @@
+// BLE device name matching the microcontroller
+const String deviceName = "PM2.5 Sensor";
+
+// BLE UUIDs matching the microcontroller
+const String serviceUUID = "91bad492-b950-4226-aa2b-4ede9fa42f59";
+const String pm1_0CharUUID = "91bad493-b950-4226-aa2b-4ede9fa42f59";
+const String pm2_5CharUUID = "91bad494-b950-4226-aa2b-4ede9fa42f59";
+const String pm10CharUUID = "91bad495-b950-4226-aa2b-4ede9fa42f59";
+const String historyCharUUID = "91bad496-b950-4226-aa2b-4ede9fa42f59";
+const String timeSyncCharUUID = "91bad497-b950-4226-aa2b-4ede9fa42f59";
